+++ packages/frontend/src/features/metricsCatalog/components/visualization/MetricsVisualization.tsx
@@ -1,18 +1,19 @@
 import {
-    assertUnreachable,
-    getFieldIdForDateDimension,
-    getItemId,
-    getMetricExplorerDataPoints,
-    getMetricExplorerDataPointsWithCompare,
-    isDimension,
-    MetricExplorerComparison,
+    capitalize,
     type MetricExplorerComparisonType,
     type MetricExplorerDateRange,
     type MetricsExplorerQueryResults,
     type TimeDimensionConfig,
     type TimeFrames,
 } from '@lightdash/common';
-import { Button, Group, Stack, Text, useMantineTheme } from '@mantine/core';
+import {
+    Button,
+    Flex,
+    Group,
+    Stack,
+    Text,
+    useMantineTheme,
+} from '@mantine/core';
 import { IconZoomReset } from '@tabler/icons-react';
 import { scaleTime } from 'd3-scale';
 import {
@@ -40,6 +41,7 @@ import {
 import MantineIcon from '../../../../components/common/MantineIcon';
 import { MetricPeekDatePicker } from '../MetricPeekDatePicker';
 import { MetricsVisualizationEmptyState } from '../MetricsVisualizationEmptyState';
+import { TimeDimensionPicker } from './TimeDimensionPicker';
 import { FORMATS } from './types';
 import { useChartZoom } from './useChartZoom';
 
@@ -78,122 +80,21 @@ type Props = {
 const MetricsVisualization: FC<Props> = ({
     results,
     dateRange,
-    comparison,
     onDateRangeChange,
     showTimeDimensionIntervalPicker,
     timeDimensionBaseField,
     setTimeDimensionOverride,
     onTimeIntervalChange,
-    isFetching,
 }) => {
     const { colors, radius, shadows, fontSizes } = useMantineTheme();
 
-    const dataPoints = useMemo(() => {
-        if (isFetching) return null;
-        if (!results?.rows) return null;
-
-        const timeDimension = results.metric.timeDimension;
-        if (!timeDimension) return null;
-
-        const dimensionId = getFieldIdForDateDimension(
-            getItemId({
-                name: timeDimension.field,
-                table: timeDimension.table,
-            }),
-            timeDimension.interval,
-        );
-
-        const dimension = results.fields[dimensionId];
-        if (!dimension || !isDimension(dimension)) return null;
-
-        switch (comparison.type) {
-            case MetricExplorerComparison.NONE:
-                return getMetricExplorerDataPoints(
-                    dimension,
-                    results.metric,
-                    results.rows,
-                );
-            case MetricExplorerComparison.PREVIOUS_PERIOD:
-                if (isFetching) {
-                    return null;
-                }
-
-                if (!results.comparisonRows) {
-                    throw new Error(
-                        `Comparison rows are required for comparison type ${comparison.type}`,
-                    );
-                }
-
-                return getMetricExplorerDataPointsWithCompare(
-                    dimension,
-                    dimension,
-                    results.metric,
-                    results.rows,
-                    results.comparisonRows,
-                    comparison,
-                );
-            case MetricExplorerComparison.DIFFERENT_METRIC:
-                if (isFetching) {
-                    return null;
-                }
-
-                if (!results.comparisonRows) {
-                    return null;
-                }
-
-                if (!results.comparisonMetric) {
-                    throw new Error(
-                        `Comparison metric not found for comparison type ${comparison.type}`,
-                    );
-                }
-
-                const compareMetric = results.comparisonMetric;
-                const compareTimeDimension = compareMetric.timeDimension;
-
-                if (!compareTimeDimension) {
-                    throw new Error(
-                        `Comparison metric should always have an associated time dimension`,
-                    );
-                }
-
-                const compareDimensionId = getFieldIdForDateDimension(
-                    getItemId({
-                        name: compareTimeDimension.field,
-                        table: compareTimeDimension.table,
-                    }),
-                    timeDimension.interval,
-                );
-
-                const compareDimension = results.fields[compareDimensionId];
-
-                if (!compareDimension || !isDimension(compareDimension)) {
-                    throw new Error(
-                        `Comparison dimension not found for comparison type ${comparison.type}`,
-                    );
-                }
-
-                return getMetricExplorerDataPointsWithCompare(
-                    dimension,
-                    compareDimension,
-                    results.metric,
-                    results.rows,
-                    results.comparisonRows,
-                    comparison,
-                );
-            default:
-                return assertUnreachable(comparison, `Unknown comparison type`);
-        }
-    }, [comparison, results, isFetching]);
-
     const data = useMemo(() => {
-        if (!dataPoints) return [];
-
-        return dataPoints
-            .map((row) => ({ ...row, dateValue: row.date.valueOf() }))
-            .sort((a, b) => a.dateValue - b.dateValue);
-    }, [dataPoints]);
+        if (!results?.results) return [];
+        return results.results;
+    }, [results]);
 
     const {
+        activeData,
         zoomState,
         handlers: {
             handleMouseDown,
@@ -201,7 +102,6 @@ const MetricsVisualization: FC<Props> = ({
             handleMouseUp,
             resetZoom,
         },
-        activeData,
     } = useChartZoom({ data });
 
     useEffect(() => {
@@ -209,33 +109,25 @@ const MetricsVisualization: FC<Props> = ({
     }, [data, resetZoom]);
 
     const xAxisConfig = useMemo(() => {
-        if (!data) return null;
-
         const timeValues = activeData.map((row) => row.dateValue);
         const timeScale = scaleTime().domain([
-            Math.min(...timeValues),
-            Math.max(...timeValues),
+            new Date(Math.min(...timeValues)),
+            new Date(Math.max(...timeValues)),
         ]);
 
         return {
-            domain: timeScale.domain().map((date) => date.valueOf()),
+            domain: [Math.min(...timeValues), Math.max(...timeValues)],
             scale: timeScale,
             type: 'number' as const,
-            ticks: timeScale.ticks(5).map((date) => date.valueOf()),
+            ticks: timeScale.ticks(5).map((d) => d.valueOf()),
             tickFormatter,
         };
-    }, [data, activeData]);
+    }, [activeData]);
 
-    const showEmptyState = dataPoints?.length === 0;
+    const showEmptyState = activeData.length === 0;
 
     return (
-        <Stack
-            spacing={showEmptyState ? 'sm' : 'lg'}
-            pb={showEmptyState ? 'sm' : undefined}
-            w="100%"
-            h="100%"
-            sx={{ flexGrow: 1 }}
-        >
+        <Stack spacing="sm" pb="sm" w="100%" h="100%">
             <Group spacing="sm" noWrap>
                 {dateRange && results?.metric.timeDimension && (
                     <MetricPeekDatePicker
@@ -266,98 +158,122 @@ const MetricsVisualization: FC<Props> = ({
             {showEmptyState && <MetricsVisualizationEmptyState />}
 
             {!showEmptyState && results && (
-                <ResponsiveContainer width="100%" height="100%">
-                    <LineChart
-                        data={activeData}
-                        margin={{
-                            top: 40,
-                            right: 40,
-                            bottom: 40,
-                            left: 40,
-                        }}
-                        onMouseDown={handleMouseDown}
-                        onMouseMove={handleMouseMove}
-                        onMouseUp={handleMouseUp}
-                    >
-                        <CartesianGrid
-                            horizontal
-                            vertical={false}
-                            stroke={colors.gray[2]}
-                            strokeDasharray="4 3"
-                        />
-
-                        <YAxis
-                            axisLine={false}
-                            tickLine={false}
-                            fontSize={11}
-                            width={4}
-                            domain={['dataMin - 1', 'dataMax + 1']}
-                            allowDataOverflow={false}
-                        />
-
-                        <XAxis
-                            dataKey="dateValue"
-                            {...xAxisConfig}
-                            axisLine={{ stroke: colors.gray[2] }}
-                            tickLine={false}
-                            fontSize={11}
-                        />
-
-                        <RechartsTooltip
-                            formatter={(value) => [value, results.metric.label]}
-                            labelFormatter={(label) =>
-                                dayjs(label).format('MMM D, YYYY')
-                            }
-                            contentStyle={{
-                                fontSize: fontSizes.xs,
-                                backgroundColor: colors.offWhite[0],
-                                borderRadius: radius.md,
-                                border: `1px solid ${colors.gray[2]}`,
-                                boxShadow: shadows.sm,
+                <Flex sx={{ flex: 1 }}>
+                    <ResponsiveContainer width="100%" height="100%">
+                        <LineChart
+                            data={activeData}
+                            margin={{
+                                right: 40,
+                                left: 40,
+                                top: 10,
                             }}
-                        />
+                            onMouseDown={handleMouseDown}
+                            onMouseMove={handleMouseMove}
+                            onMouseUp={handleMouseUp}
+                        >
+                            <Legend
+                                verticalAlign="top"
+                                height={50}
+                                margin={{ bottom: 20 }}
+                                formatter={(value) => (
+                                    <Text span c="dark.5" size={14} fw={400}>
+                                        {value}
+                                    </Text>
+                                )}
+                            />
+                            <CartesianGrid
+                                horizontal
+                                vertical={false}
+                                stroke={colors.gray[2]}
+                                strokeDasharray="4 3"
+                            />
 
-                        <Line
-                            name={results.metric.label}
-                            type="monotone"
-                            dataKey="metric"
-                            stroke={colors.indigo[6]}
-                            strokeWidth={1.6}
-                            dot={false}
-                            legendType="plainline"
-                        />
+                            <YAxis
+                                axisLine={false}
+                                tickLine={false}
+                                fontSize={11}
+                                width={4}
+                                domain={['dataMin - 1', 'dataMax + 1']}
+                                allowDataOverflow={false}
+                            />
 
-                        {results.comparisonRows && (
+                            <XAxis
+                                dataKey="dateValue"
+                                {...xAxisConfig}
+                                axisLine={{ stroke: colors.gray[2] }}
+                                tickLine={false}
+                                fontSize={11}
+                            />
+
+                            <RechartsTooltip
+                                formatter={(value) => [
+                                    value,
+                                    results.metric.label,
+                                ]}
+                                labelFormatter={(label) =>
+                                    dayjs(label).format('MMM D, YYYY')
+                                }
+                                contentStyle={{
+                                    fontSize: fontSizes.xs,
+                                    backgroundColor: colors.offWhite[0],
+                                    borderRadius: radius.md,
+                                    border: `1px solid ${colors.gray[2]}`,
+                                    boxShadow: shadows.sm,
+                                }}
+                            />
+
                             <Line
-                                name={`${results.metric.label} (comparison)`}
-                                type="monotone"
-                                dataKey="compareMetric"
-                                stroke={colors.indigo[4]}
-                                strokeDasharray={'3 3'}
-                                strokeWidth={1.3}
+                                name={results.metric.label}
+                                type="linear"
+                                dataKey="metric"
+                                stroke={colors.indigo[6]}
+                                strokeWidth={1.6}
                                 dot={false}
                                 legendType="plainline"
                             />
-                        )}
 
-                        <Legend
-                            formatter={(value) => (
-                                <Text span c="dark.5" size={14} fw={400}>
-                                    {value}
-                                </Text>
+                            {results.compareMetric && (
+                                <Line
+                                    name={`${results.metric.label} (comparison)`}
+                                    type="linear"
+                                    dataKey="compareMetric"
+                                    stroke={colors.indigo[4]}
+                                    strokeDasharray={'3 3'}
+                                    strokeWidth={1.3}
+                                    dot={false}
+                                    legendType="plainline"
+                                />
                             )}
-                        />
 
-                        {zoomState.refAreaLeft && zoomState.refAreaRight && (
-                            <ReferenceArea
-                                x1={zoomState.refAreaLeft}
-                                x2={zoomState.refAreaRight}
-                                strokeOpacity={0.3}
-                                fill={colors.gray[3]}
-                            />
-                        )}
-                    </LineChart>
-                </ResponsiveContainer>
+                            {zoomState.refAreaLeft &&
+                                zoomState.refAreaRight && (
+                                    <ReferenceArea
+                                        x1={zoomState.refAreaLeft}
+                                        x2={zoomState.refAreaRight}
+                                        strokeOpacity={0.3}
+                                        fill={colors.gray[3]}
+                                    />
+                                )}
+                        </LineChart>
+                    </ResponsiveContainer>
+                </Flex>
+            )}
+            {results?.metric.availableTimeDimensions && (
+                <Group position="center" mt="auto">
+                    <Group align="center" noWrap>
+                        <Text fw={500} c="gray.7" fz="sm">
+                            Date ({capitalize(timeDimensionBaseField.interval)})
+                        </Text>
+
+                        <TimeDimensionPicker
+                            fields={results.metric.availableTimeDimensions}
+                            dimension={timeDimensionBaseField}
+                            onChange={(config) =>
+                                setTimeDimensionOverride(config)
+                            }
+                        />
+                    </Group>
+                </Group>
             )}
         </Stack>
     );